Move gas estimation tests to separate module
