+++ core/node/api_server/src/tx_sender/tests/gas_estimation.rs
@@ -1,257 +1,36 @@
-//! Tests for the transaction sender.
+//! Tests for gas estimation (mostly with the real oneshot VM executor).
 
-use std::{collections::HashMap, time::Duration};
+use std::collections::HashMap;
 
 use assert_matches::assert_matches;
-use test_casing::{test_casing, Product, TestCases};
-use zksync_contracts::{get_loadnext_contract, test_contracts::LoadnextContractExecutionParams};
-use zksync_multivm::interface::ExecutionResult;
-use zksync_node_fee_model::MockBatchFeeParamsProvider;
-use zksync_node_genesis::{insert_genesis_batch, GenesisParams};
-use zksync_node_test_utils::{create_l2_block, create_l2_transaction, prepare_recovery_snapshot};
+use test_casing::{test_casing, Product};
+use zksync_contracts::get_loadnext_contract;
 use zksync_system_constants::CODE_ORACLE_ADDRESS;
 use zksync_types::{
-    api,
     api::state_override::{Bytecode, OverrideAccount, OverrideState},
-    get_nonce_key,
     web3::keccak256,
-    K256PrivateKey, L1BatchNumber, L2BlockNumber, StorageLog,
+    K256PrivateKey,
 };
-use zksync_utils::{bytecode::hash_bytecode, u256_to_h256};
-use zksync_vm_executor::oneshot::MockOneshotExecutor;
+use zksync_utils::bytecode::hash_bytecode;
 
-use super::{gas_estimation::GasEstimator, *};
+use super::*;
 use crate::{
-    execution_sandbox::BlockStartInfo,
     testonly::{
         inflate_bytecode, read_counter_contract_bytecode, read_expensive_contract_bytecode,
         read_infinite_loop_contract_bytecode, read_precompiles_contract_bytecode, TestAccount,
         COUNTER_CONTRACT_ADDRESS, EXPENSIVE_CONTRACT_ADDRESS, INFINITE_LOOP_CONTRACT_ADDRESS,
         LOAD_TEST_ADDRESS, PRECOMPILES_CONTRACT_ADDRESS,
     },
-    web3::testonly::create_test_tx_sender,
+    tx_sender::gas_estimation::GasEstimator,
 };
 
 /// Initial pivot multiplier empirically sufficient for most tx types.
 const DEFAULT_MULTIPLIER: f64 = 64.0 / 63.0;
 
 #[tokio::test]
-async fn getting_nonce_for_account() {
-    let l2_chain_id = L2ChainId::default();
-    let test_address = Address::repeat_byte(1);
-    let pool = ConnectionPool::<Core>::test_pool().await;
-    let mut storage = pool.connection().await.unwrap();
-    insert_genesis_batch(&mut storage, &GenesisParams::mock())
-        .await
-        .unwrap();
-    // Manually insert a nonce for the address.
-    let nonce_key = get_nonce_key(&test_address);
-    let nonce_log = StorageLog::new_write_log(nonce_key, H256::from_low_u64_be(123));
-    storage
-        .storage_logs_dal()
-        .append_storage_logs(L2BlockNumber(0), &[nonce_log])
-        .await
-        .unwrap();
-
-    let tx_executor = MockOneshotExecutor::default();
-    let tx_executor = SandboxExecutor::mock(tx_executor).await;
-    let (tx_sender, _) = create_test_tx_sender(pool.clone(), l2_chain_id, tx_executor).await;
-
-    let nonce = tx_sender.get_expected_nonce(test_address).await.unwrap();
-    assert_eq!(nonce, Nonce(123));
-
-    // Insert another L2 block with a new nonce log.
-    storage
-        .blocks_dal()
-        .insert_l2_block(&create_l2_block(1))
-        .await
-        .unwrap();
-    let nonce_log = StorageLog {
-        value: H256::from_low_u64_be(321),
-        ..nonce_log
-    };
-    storage
-        .storage_logs_dal()
-        .insert_storage_logs(L2BlockNumber(1), &[nonce_log])
-        .await
-        .unwrap();
-
-    let nonce = tx_sender.get_expected_nonce(test_address).await.unwrap();
-    assert_eq!(nonce, Nonce(321));
-    let missing_address = Address::repeat_byte(0xff);
-    let nonce = tx_sender.get_expected_nonce(missing_address).await.unwrap();
-    assert_eq!(nonce, Nonce(0));
-}
-
-#[tokio::test]
-async fn getting_nonce_for_account_after_snapshot_recovery() {
-    const SNAPSHOT_L2_BLOCK_NUMBER: L2BlockNumber = L2BlockNumber(42);
-
-    let pool = ConnectionPool::<Core>::test_pool().await;
-    let mut storage = pool.connection().await.unwrap();
-    let test_address = Address::repeat_byte(1);
-    let other_address = Address::repeat_byte(2);
-    let nonce_logs = [
-        StorageLog::new_write_log(get_nonce_key(&test_address), H256::from_low_u64_be(123)),
-        StorageLog::new_write_log(get_nonce_key(&other_address), H256::from_low_u64_be(25)),
-    ];
-    prepare_recovery_snapshot(
-        &mut storage,
-        L1BatchNumber(23),
-        SNAPSHOT_L2_BLOCK_NUMBER,
-        &nonce_logs,
-    )
-    .await;
-
-    let l2_chain_id = L2ChainId::default();
-    let tx_executor = MockOneshotExecutor::default();
-    let tx_executor = SandboxExecutor::mock(tx_executor).await;
-    let (tx_sender, _) = create_test_tx_sender(pool.clone(), l2_chain_id, tx_executor).await;
-
-    storage
-        .blocks_dal()
-        .insert_l2_block(&create_l2_block(SNAPSHOT_L2_BLOCK_NUMBER.0 + 1))
-        .await
-        .unwrap();
-    let new_nonce_logs = vec![StorageLog::new_write_log(
-        get_nonce_key(&test_address),
-        H256::from_low_u64_be(321),
-    )];
-    storage
-        .storage_logs_dal()
-        .insert_storage_logs(SNAPSHOT_L2_BLOCK_NUMBER + 1, &new_nonce_logs)
-        .await
-        .unwrap();
-
-    let nonce = tx_sender.get_expected_nonce(test_address).await.unwrap();
-    assert_eq!(nonce, Nonce(321));
-    let nonce = tx_sender.get_expected_nonce(other_address).await.unwrap();
-    assert_eq!(nonce, Nonce(25));
-    let missing_address = Address::repeat_byte(0xff);
-    let nonce = tx_sender.get_expected_nonce(missing_address).await.unwrap();
-    assert_eq!(nonce, Nonce(0));
-}
-
-#[tokio::test]
-async fn submitting_tx_requires_one_connection() {
-    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
-    let mut storage = pool.connection().await.unwrap();
-    insert_genesis_batch(&mut storage, &GenesisParams::mock())
-        .await
-        .unwrap();
-
-    let l2_chain_id = L2ChainId::default();
-    let fee_input = MockBatchFeeParamsProvider::default()
-        .get_batch_fee_input_scaled(1.0, 1.0)
-        .await
-        .unwrap();
-    let (base_fee, gas_per_pubdata) =
-        derive_base_fee_and_gas_per_pubdata(fee_input, ProtocolVersionId::latest().into());
-    let tx = create_l2_transaction(base_fee, gas_per_pubdata);
-    let tx_hash = tx.hash();
-
-    // Manually set sufficient balance for the tx initiator.
-    let balance_key = storage_key_for_eth_balance(&tx.initiator_account());
-    let storage_log = StorageLog::new_write_log(balance_key, u256_to_h256(U256::one() << 64));
-    storage
-        .storage_logs_dal()
-        .append_storage_logs(L2BlockNumber(0), &[storage_log])
-        .await
-        .unwrap();
-    drop(storage);
-
-    let mut tx_executor = MockOneshotExecutor::default();
-    tx_executor.set_tx_responses(move |received_tx, _| {
-        assert_eq!(received_tx.hash(), tx_hash);
-        ExecutionResult::Success { output: vec![] }
-    });
-    let tx_executor = SandboxExecutor::mock(tx_executor).await;
-    let (tx_sender, _) = create_test_tx_sender(pool.clone(), l2_chain_id, tx_executor).await;
-
-    let submission_result = tx_sender.submit_tx(tx).await.unwrap();
-    assert_matches!(submission_result.0, L2TxSubmissionResult::Added);
-
-    let mut storage = pool.connection().await.unwrap();
-    storage
-        .transactions_web3_dal()
-        .get_transaction_by_hash(tx_hash, l2_chain_id)
-        .await
-        .unwrap()
-        .expect("transaction is not persisted");
-}
-
-#[tokio::test]
-async fn eth_call_requires_single_connection() {
-    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
-    let mut storage = pool.connection().await.unwrap();
-    let genesis_params = GenesisParams::mock();
-    insert_genesis_batch(&mut storage, &genesis_params)
-        .await
-        .unwrap();
-    let start_info = BlockStartInfo::new(&mut storage, Duration::MAX)
-        .await
-        .unwrap();
-    let block_id = api::BlockId::Number(api::BlockNumber::Latest);
-    let block_args = BlockArgs::new(&mut storage, block_id, &start_info)
-        .await
-        .unwrap();
-    drop(storage);
-
-    let tx = create_l2_transaction(10, 100);
-    let tx_hash = tx.hash();
-
-    let mut tx_executor = MockOneshotExecutor::default();
-    tx_executor.set_call_responses(move |received_tx, _| {
-        assert_eq!(received_tx.hash(), tx_hash);
-        ExecutionResult::Success {
-            output: b"success!".to_vec(),
-        }
-    });
-    let tx_executor = SandboxExecutor::mock(tx_executor).await;
-    let (tx_sender, _) = create_test_tx_sender(
-        pool.clone(),
-        genesis_params.config().l2_chain_id,
-        tx_executor,
-    )
-    .await;
-    let call_overrides = CallOverrides {
-        enforced_base_fee: None,
-    };
-    let output = tx_sender
-        .eth_call(block_args, call_overrides, tx, None)
-        .await
-        .unwrap();
-    assert_eq!(output, b"success!");
-}
-
-async fn create_real_tx_sender() -> TxSender {
-    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
-    let mut storage = pool.connection().await.unwrap();
-    let genesis_params = GenesisParams::mock();
-    insert_genesis_batch(&mut storage, &genesis_params)
-        .await
-        .unwrap();
-    drop(storage);
-
-    let genesis_config = genesis_params.config();
-    let executor_options = SandboxExecutorOptions::new(
-        genesis_config.l2_chain_id,
-        AccountTreeId::new(genesis_config.fee_account),
-        u32::MAX,
-    )
-    .await
-    .unwrap();
-
-    let pg_caches = PostgresStorageCaches::new(1, 1);
-    let tx_executor = SandboxExecutor::real(executor_options, pg_caches, usize::MAX);
-    create_test_tx_sender(pool, genesis_params.config().l2_chain_id, tx_executor)
-        .await
-        .0
-}
-
-#[tokio::test]
 async fn initial_gas_estimation_is_somewhat_accurate() {
-    let tx_sender = create_real_tx_sender().await;
+    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
+    let tx_sender = create_real_tx_sender(pool).await;
 
     let alice = K256PrivateKey::random();
     let transfer_value = U256::from(1_000_000_000);
@@ -260,8 +39,7 @@ async fn initial_gas_estimation_is_somewhat_accurate() {
         ..OverrideAccount::default()
     };
     let state_override = StateOverride::new(HashMap::from([(alice.address(), account_overrides)]));
-    // fee params don't matter; we adjust via `adjust_transaction_fee()`
-    let tx = alice.create_transfer(transfer_value, 55, 555);
+    let tx = alice.create_transfer(transfer_value);
 
     let mut estimator = GasEstimator::new(&tx_sender, tx.into(), Some(state_override))
         .await
@@ -289,31 +67,6 @@ async fn initial_gas_estimation_is_somewhat_accurate() {
     let (vm_result, _) = estimator.unadjusted_step(initial_pivot).await.unwrap();
     assert!(!vm_result.result.is_failed(), "{:?}", vm_result.result);
 }
-
-const LOAD_TEST_CASES: TestCases<LoadnextContractExecutionParams> = test_casing::cases! {[
-    LoadnextContractExecutionParams::default(),
-    // No storage modification
-    LoadnextContractExecutionParams {
-        writes: 0,
-        events: 0,
-        ..LoadnextContractExecutionParams::default()
-    },
-    // Moderately deep recursion (very deep recursion is tested separately)
-    LoadnextContractExecutionParams {
-        recursive_calls: 10,
-        ..LoadnextContractExecutionParams::default()
-    },
-    // No deploys
-    LoadnextContractExecutionParams {
-        deploys: 0,
-        ..LoadnextContractExecutionParams::default()
-    },
-    // Lots of deploys
-    LoadnextContractExecutionParams {
-        deploys: 10,
-        ..LoadnextContractExecutionParams::default()
-    },
-]};
 
 #[test_casing(5, LOAD_TEST_CASES)]
 #[tokio::test]
@@ -400,7 +153,8 @@ async fn test_initial_estimate(
     tx: L2Tx,
     initial_pivot_multiplier: f64,
 ) -> VmExecutionResultAndLogs {
-    let tx_sender = create_real_tx_sender().await;
+    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
+    let tx_sender = create_real_tx_sender(pool).await;
     let mut estimator = GasEstimator::new(&tx_sender, tx.into(), Some(state_override))
         .await
         .unwrap();
@@ -421,7 +175,8 @@ async fn test_initial_estimate(
 }
 
 async fn test_initial_estimate_error(state_override: StateOverride, tx: L2Tx) -> SubmitTxError {
-    let tx_sender = create_real_tx_sender().await;
+    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
+    let tx_sender = create_real_tx_sender(pool).await;
     let mut estimator = GasEstimator::new(&tx_sender, tx.into(), Some(state_override))
         .await
         .unwrap();
@@ -561,7 +316,7 @@ async fn revert_during_initial_estimate() {
         contract_overrides,
     )]));
 
-    let tx = alice.create_reverting_counter_tx();
+    let tx = alice.create_counter_tx(1.into(), true);
     let err = test_initial_estimate_error(state_override, tx).await;
     let SubmitTxError::ExecutionReverted(err, _) = err else {
         panic!("Unexpected error: {err:?}");
@@ -590,12 +345,11 @@ async fn out_of_gas_during_initial_estimate() {
 
 #[tokio::test]
 async fn insufficient_funds_error_for_transfer() {
-    let tx_sender = create_real_tx_sender().await;
+    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
+    let tx_sender = create_real_tx_sender(pool).await;
 
     let alice = K256PrivateKey::random();
-    let transfer_value = 1_000_000_000.into();
-    // fee params don't matter; they should be overwritten by the estimation logic
-    let tx = alice.create_transfer(transfer_value, 55, 555);
+    let tx = alice.create_transfer(1_000_000_000.into());
     let fee_scale_factor = 1.0;
     // Without overrides, the transaction should fail because of insufficient balance.
     let err = tx_sender
@@ -616,7 +370,8 @@ async fn test_estimating_gas(
     tx: L2Tx,
     acceptable_overestimation: u64,
 ) {
-    let tx_sender = create_real_tx_sender().await;
+    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
+    let tx_sender = create_real_tx_sender(pool).await;
 
     let fee_scale_factor = 1.0;
     let fee = tx_sender
@@ -665,8 +420,7 @@ async fn estimating_gas_for_transfer(acceptable_overestimation: u64) {
         ..OverrideAccount::default()
     };
     let state_override = StateOverride::new(HashMap::from([(alice.address(), account_overrides)]));
-    // fee params don't matter; they should be overwritten by the estimation logic
-    let tx = alice.create_transfer(transfer_value, 55, 555);
+    let tx = alice.create_transfer(transfer_value);
 
     test_estimating_gas(state_override, tx, acceptable_overestimation).await;
 }
@@ -751,8 +505,9 @@ async fn estimating_gas_for_reverting_tx() {
         contract_overrides,
     )]));
 
-    let tx = alice.create_reverting_counter_tx();
-    let tx_sender = create_real_tx_sender().await;
+    let tx = alice.create_counter_tx(1.into(), true);
+    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
+    let tx_sender = create_real_tx_sender(pool).await;
 
     let fee_scale_factor = 1.0;
     let acceptable_overestimation = 0;
@@ -785,7 +540,8 @@ async fn estimating_gas_for_infinite_loop_tx() {
     )]));
 
     let tx = alice.create_infinite_loop_tx();
-    let tx_sender = create_real_tx_sender().await;
+    let pool = ConnectionPool::<Core>::constrained_test_pool(1).await;
+    let tx_sender = create_real_tx_sender(pool).await;
 
     let fee_scale_factor = 1.0;
     let acceptable_overestimation = 0;
